Move files from viewer app to renderer-V2 lib.
