+++ src/apps/viewer/viewer/CMakeLists.txt
@@ -6,12 +6,9 @@ configure_file(build_info.h.in build_info.h @ONLY)
 
 set(${TARGET_NAME}_HEADERS
     CameraSystem.h
-    ColorPalettes.h
     DebugAndLogGui.h
-    DebugDrawUtilities.h
     DrawQuad.h
     EnvironmentUtilities.h
-    FrustumUtilities.h
     GraphCommon.h
     GraphGui.h
     Json.h
