+++ src/libs/snac-renderer-V2/snac-renderer-V2/utilities/VertexStreamUtilities.cpp
