+++ src/libs/snac-renderer-V2/snac-renderer-V2/utilities/VertexStreamUtilities.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Model.h"
+#include "../Model.h"
 
 
 namespace ad::renderer {
